pyproject.toml
[tool.poetry]
name = "workshop-day2"
version = "0.1.0"
description = ""
authors = ["Breno Mendes <b.mendesbernardo2020@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
gdown = "^5.1.0"
duckdb = "^0.10.0"
streamlit = "^1.32.1"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
sqlalchemy = "^2.0.28"
load-dotenv = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
